move e2e tests, split up maindockerbuild
